Initial commit: NeoCalc
NeoCalc started earlier this year as InterCalc while I was teaching Python to kids as part of their curriculum. The name that first was a reference to the school's name
became as I forked it to improve it on my own the NeoCalc, a name that represent a "New" calculator and also a play with my first name Nilton, which the first two letters
are pronounced in my native language, portuguese, just like "neo".

Over the past few months, I also grew increasingly frustrated with the standard GNOME Calculator, while trying to calculate my rent
and other expenses, because every single time I had to open a new Calculator instance which is frustrating, motivating me to experiment with building my own.

This project is primarily a learning exercise. The core calculation logic is written in Rust, a language I am still trying to grasp, while Python and CSS are used for the UI and design layer, allowing faster iteration and experimentation.
NeoCalc is not intended as a replacement for existing calculator apps. It was built for my own use case and as a multi-language programming exercise. That said, the code is open to improvement, and others are welcome to extend it into something useful for a broader audience.

Signed-off-by: Nilton Perim Neto <niltonperimneto@gmail.com>